PyodideRuntime should be one of the runtimes
